+++ example/lib/infinite_grouped_list/infinite_grouped_list.dart
@@ -533,17 +533,20 @@ class InfiniteGroupedListState<Cell, GroupBy, Group> extends State<InfiniteGroup
                       return MapEntry(
                         index,
                         SliverStickyCollapsablePanel(
-                          iOSStyleSticky: false,
+                          key: Key("SliverStickyCollapsablePanel_$index"),
                           scrollController: _scrollController,
+                          panelController: StickyCollapsablePanelController(
+                            key: title.toString(),
+                            disableCollapsable: (index % 2) == 1,
+                          ),
                           paddingAfterCollapse: const EdgeInsets.only(bottom: 10),
-                          controller: StickyCollapsablePanelController(key: title.toString()),
                           sticky: widget.stickyGroups,
-                          disableCollapsable: (index % 2) == 1,
                           expandCallback: (isExpanded) {
                             if (!isExpanded) {
                               _loadMore();
                             }
                           },
+                          iOSStyleSticky: false,
                           headerBuilder: (context, status) {
                             return widget.groupTitleBuilder(
                               index,
@@ -554,6 +557,7 @@ class InfiniteGroupedListState<Cell, GroupBy, Group> extends State<InfiniteGroup
                               status.scrollPercentage,
                             );
                           },
+                          panelAnimationDuration: Duration(milliseconds: 300),
                           headerSize: Size(MediaQuery.of(context).size.width, 50),
                           sliverPanel: widget.listStyle == ListStyle.listView
                               ? SliverList(