+++ lib/utils/utils.dart
@@ -1,3 +1,2 @@
-export 'sliver_sticky_collapsable_panel_controller.dart';
 export 'sliver_sticky_collapsable_panel_status.dart';
 export 'value_layout_builder.dart';